==> 03_merge_fitness.Rmd <==
<!--pandoc
t: html
default-image-extension: png
template: templates/html.template
toc-depth: 4
toc:

t: latex
s:
S:
latex-engine: xelatex
template: templates/latex.template
default-image-extension: pdf
toc-depth: 4
toc:
-->

```{r echo=FALSE, cache=FALSE, results="hide"}
opts_chunk$set(cache=TRUE, fig.cap='', dev=c('png', 'pdf'), fig.width=6,
  fig.height=4)
opts_knit$set('upload.fun' = function(x) sub("\\.[[:alnum:]]*$", "", x) ) 
opts_chunk$set(cache.path='cache/03/')
```
# Fitness and Standardization

###Load Data

```{r fitness_data}
library(MASS) # MASS clashes with dplyr... so always load first
library(pander) # pander clashes with dplyr... so always load first
library(dplyr)
set.alignment('right', row.names = 'left')
load("data/analyses_data/raneffs_blups.RData")
# load("data/analyses_data/pca.RData")
fitness <- read.table(file = "data/fitness+competition.csv", sep = ',',
  header = TRUE, stringsAsFactors = FALSE)
fitness <- tbl_df(fitness)
```

## Merge fitness with blups

```{r merge_fitness }
fitness$ID <- as.character(fitness$ID)
fit_raneff_data <- inner_join(fitness, raneffs_blups, by = "ID")
```

## Relative Fitness
Calculate relative fitness for each year & population combination. Two populations (Grids).  
Three measures of fitness:

1. ars_all = Annual reproductive success over all litters (no. pups that survived overwinter)  
2. kprod   = Fecundity (kids produced)  
3. prop    = Offpsring overwinter survival (proportion of pups produced that survived overwinter)  

```{r RelativeFitness }
# Calculate offspring overwitner survival
fit_raneff_data <- fit_raneff_data %>% mutate(prop = ars_all/kprod)

fit_raneff_data <- fit_raneff_data %>%
  group_by(Grid, Year, itt) %>%
  mutate(rel_ars = ars_all / mean(ars_all),
   rel_kpd = kprod / mean(kprod),
   rel_ows = prop / mean(prop)
   )
```
Now:

* rel_ars = relative ARS
* rel_kpd = relative fecunidty
* felOWS = relative offspring overwinter survival

## Standardize Variables
Standaradized to mean 0 and sd 1. Standardized variables renamed from xxx to xxx_s or xxx_sy (for standardized within year). Standardized within each BLUP itteration.

```{r standardize-traits}
# Standardize within itteration and year
fit_raneff_data <- fit_raneff_data %>%
  group_by(itt, Year, add = FALSE) %>%
  mutate(
    aggression_sy  = (aggression  - mean(aggression, na.rm = TRUE))  /
      sd(aggression, na.rm = TRUE),
    activity_sy    = (activity    - mean(activity, na.rm = TRUE))    /
      sd(activity, na.rm = TRUE),
    docility_sy    = (docility    - mean(docility, na.rm = TRUE))    /
      sd(docility, na.rm = TRUE),
    competition_sy = (competition - mean(competition, na.rm = TRUE)) /
      sd(competition, na.rm = TRUE)
      )
# Standardize within itteration
fit_raneff_data <- fit_raneff_data %>%
  group_by(itt, add = FALSE) %>%
  mutate(
    aggression_s  = (aggression  - mean(aggression, na.rm = TRUE))  /
      sd(aggression, na.rm = TRUE),
    activity_s    = (activity    - mean(activity, na.rm = TRUE))    /
      sd(activity, na.rm = TRUE),
    docility_s    = (docility    - mean(docility, na.rm = TRUE))    /
      sd(docility, na.rm = TRUE),
    competition_s = (competition - mean(competition, na.rm = TRUE)) /
      sd(competition, na.rm = TRUE)
      )
```
```{r test_stadardization, results='asis'}
fit_raneff_data %>%
  group_by(itt, add = FALSE) %>%
  summarise(
    v_agg = var(aggression_s, na.rm = TRUE),
    v_act = var(activity_s, na.rm = TRUE),
    v_doc = var(docility_s, na.rm = TRUE)
    ) %>%
  head(.,n=10) %>%
  pandoc.table(.)
```
```{r test_stadardization_gy, results='asis'}
fit_raneff_data %>%
  group_by(itt, Year, add = FALSE) %>%
  summarise(
    v_agg = var(aggression_sy, na.rm = TRUE),
    v_act = var(activity_sy, na.rm = TRUE),
    v_doc = var(docility_sy, na.rm = TRUE)
    ) %>%
  head(.,n=10) %>%
  pandoc.table(.)
```

## Sample Sizes
```{r samples, results='asis'}
fit_raneff_data %>%
  filter(itt == "1") %>%
  group_by(Grid, Year, add = FALSE) %>%
  summarise(n()) %>%
  pandoc.table(.)
  
fit_raneff_data %>%
filter(itt == "1") %>%
  group_by(Year, add = FALSE) %>%
  summarise(n()) %>%
  pandoc.table(.)
```

```{r save_data }
save(fit_raneff_data, file = "data/analyses_data/fit_raneff_data.RData")
```
